Fix gemspec
The name of the gem needs to match the name of main file. Otherwise, bundler won't find any file to require when
executing "Bundler.require(*Rails.groups)" causing the app to fail due to NameError.
A workaround was to add the attribute "require: 'yaaf'" to the Gemfile, but that's not a common thing to do.