Adjust specs for testing callback exceptions
This commit removes the previous specs/support files testing callback exceptions, and
replaces them with a very explicit example of raising specific
exceptions in callbacks.

This ensures that each type of exception that can be rescued during `#save`
is covered by tests, and also adds a test scenario for a non-rescueable
exception.