Adjust #save/! for callback validation exceptions (#67)
* Adjust #save/! for callback validation exceptions

In digging into how YAAF works under the hood, and looking at our own
use-cases, we came across an instance where - in an `after_save`
callback - we would need to have the ability to trigger a rollback of
the entire callback chain and transaction, and still return `false` for
use in the controller flow.

After digging through how typical Rails models handle this - we noticed
that
[`ActiveRecord::Persistence.save`](https://github.com/rails/rails/blob/83217025a171593547d1268651b446d3533e2019/activerecord/lib/active_record/persistence.rb#L473-L477)
calls the same method as
[`ActiveRecord::Persistence.save!`](https://github.com/rails/rails/blob/83217025a171593547d1268651b446d3533e2019/activerecord/lib/active_record/persistence.rb#L506-L508)
under the hood - it just rescues from `ActiveRecord::RecordInvalid`.

This would seem that, instead of just checking for `valid?` - it is
instead calling `validate!`, and rescuing from the exception that pops
up given validation errors.

This commit adjusts the `Form#save` and `Form#save!` implementations to
mimick how `ActiveRecord::Persistence` is handling the same problem,
allowing for validation errors in callbacks to still trigger the
rollback and but not break the control flow when calling `form.save`.

It also rescues from two other exceptions:

1. `ActiveModel::ValidationError`, as that is what is raised when
   validations fail on an `ActiveModel::Model` object.
2. `ActiveRecord::RecordNotSaved`, as that is what is raised if
   [`create_or_update` returns `false`](https://github.com/rails/rails/blob/83217025a171593547d1268651b446d3533e2019/activerecord/lib/active_record/persistence.rb#L507)

---

Something to note with this change: `save!` has been adjusted to call
the underlying persistence method, `save_form`, which either returns
`true`, or raises exceptions in the event of validation errors. This
change means that raising `ActiveRecord::RecordNotSaved` will
effectively never happen.

I made this change so that the validation errors will still be output
in the exception message (just like `ActiveRecord::Model#save!` would).

In looking at [how Rails handles this](https://github.com/rails/rails/blob/83217025a171593547d1268651b446d3533e2019/activerecord/lib/active_record/persistence.rb#L899),
it is handling a few more scenarios that don't fit within the current
implementation of `YAAF::Form`, ex: returning `false` if `destroyed?`.

All this to say, the `|| raise(ActiveRecord::RecordNotSaved` can be
removed given it will never be reached from any code path.

* Remove unreachable raise from #save!

With how `#save_form` currently works, it is impossible for it to ever
return anything but `true`.

As such, the `raise(ActiveRecord::RecordNotSaved` will never be called.

This commit removes it.

* Adjust specs for testing callback exceptions

This commit removes the previous specs/support files testing callback exceptions, and
replaces them with a very explicit example of raising specific
exceptions in callbacks.

This ensures that each type of exception that can be rescued during `#save`
is covered by tests, and also adds a test scenario for a non-rescueable
exception.