Delegate #transaction to ActiveRecord::Base
For our project, we're using some custom database connection
logic/pooling, along with multiple databases.

Because of this - it means that we're no longer able to reach for
`::ActiveRecord::Base` to wrap transactions.

Instead, we have to use separate classes, ex:

```ruby
ClientApplicationRecord.transaction do
  # do stuff with client connections
end

ApplicationRecord.transaction do
  # do stuff with application connections
end
```

Delegating `#transaction` to `::ActiveRecord::Base` allows us to safely
override the delegation for Client specific forms.

```ruby
class ClientForm < YAAF::Form
  delegate :transaction, to: ClientApplicationRecord
end
```

as opposed to having to override the entire `save_in_transaction`
method:

```ruby
class ClientForm < YAAF::Form
  # lots of duplicated logic that may change in future gem versions
  def save_in_transaction(options)
    ::ClientApplicationRecord.transaction do
      run_callbacks :save do
        save_models(options)
      end
    end
  rescue => e
    handle_transaction_rollback(e)
  end
end
```

I went back and forth on whether or not to just make a private
`#database_connection_class` method, but doing so could mean that a
future committer refactor/remove the change. Making it public means that
it can be tested (as it's poor form to directly test private methods),
which adds a bit of friction for removal.

As for a _public_ `#database_connection_class` method, `Form`'s
_written_ public API is already so small, that having this method hang
out in the file with `#save` and `#save!` just didn't quite feel right.